YATA-38 Refactoring backend
