+++ yata-server/build.gradle
@@ -13,7 +13,6 @@ dependencies {
     compileOnly ('org.projectlombok:lombok')
     annotationProcessor ('org.projectlombok:lombok')
     compile ('org.apache.commons:commons-lang3:3.11')
-
 }
 
 bootJar.enabled = true
