+++ yata-server/src/main/java/project/yata/web/AccountController.java
@@ -21,13 +21,7 @@ public class AccountController {
 
     @GetMapping("")
     public ResponseEntity<ApiResponse> getAccount(@RequestParam String email) {
-        return ResponseEntity
-                .status(HttpStatus.OK)
-                .body(
-                        ApiResponse.success(
-                                accountService.getAccount(email)
-                        )
-                );
+        return ResponseEntity.status(HttpStatus.OK)
+                .body(ApiResponse.success(accountService.getAccount(email)));
     }
-
 }
