+++ yata-server/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.profiles.active=local
+spring.profiles.active=dev
 
 jwt.access=15
 jwt.refresh=60
@@ -8,4 +8,3 @@ server.servlet.encoding.force=true
 server.servlet.encoding.force-response=true
 server.servlet.encoding.force-request=true
 server.servlet.encoding.charset=UTF-8
-#spring.jpa.open-in-view=true
